+++ qualitycheck-20190115/alibabacloud_qualitycheck20190115/models/__init__.py
@@ -321,6 +321,7 @@ from ._graph_flow_node import GraphFlowNodeNextNodes
 from ._graph_flow_node import GraphFlowNodeProperties
 from ._judge_node_meta_desc import JudgeNodeMetaDescExpressionMetaDesc
 from ._next_node_situations import NextNodeSituationsConditionGroup
+from ._operator_basic_info import OperatorBasicInfoParamDimensions
 from ._operator_basic_info import OperatorBasicInfoParam
 from ._rule_test_dialogue import RuleTestDialogueContent
 from ._scheme_check_type import SchemeCheckTypeSchemeScoreInfoList
@@ -928,6 +929,7 @@ __all__ = [
     GraphFlowNodeProperties,
     JudgeNodeMetaDescExpressionMetaDesc,
     NextNodeSituationsConditionGroup,
+    OperatorBasicInfoParamDimensions,
     OperatorBasicInfoParam,
     RuleTestDialogueContent,
     SchemeCheckTypeSchemeScoreInfoList,